+++ packages/react/src/components/Calendar/src/calendar.types.ts
@@ -1,19 +1,19 @@
 import { ComponentProps } from "react";
 import { DayPicker } from "react-day-picker";
-import { BoxProps } from "@nimbus-ds/components";
+import { BoxProperties, BoxProps } from "@nimbus-ds/components";
 
 type CalendarExtends = ComponentProps<typeof DayPicker>;
 
-export type CalendarProps = CalendarExtends & {
+export interface CalendarProperties {
   /**
    * Ability to hide the border of the calendar container. Useful for including the calendar inside other components such as Modal, Popover or Card.
    * @default false
    */
   hideBorder?: boolean;
   /**
-   * Spreads the properties of the Box component to the calendar wrapper.
+   * Spreads the properties of the Box component to thˇe calendar wrapper.
    */
-  containerProps?: BoxProps;
+  containerProps?: BoxProperties;
   /**
    * If true will stick the weekday indicators to the top of the component. Useful for when creating scrolling calendars with a display of >1 months. Only works when property numberOfMonths is set to a number greater than 1.
    * @default false
@@ -24,4 +24,12 @@ export type CalendarProps = CalendarExtends & {
    * @default false
    */
   fullWidthDays?: boolean;
-};
+}
+
+export type CalendarProps = CalendarProperties &
+  CalendarExtends & {
+    /**
+     * Spreads the properties of the Box component to thˇe calendar wrapper.
+     */
+    containerProps?: BoxProps;
+  };