+++ packages/react/src/components/MenuButton/src/menuButton.stories.tsx
@@ -1,6 +1,6 @@
 import React, { forwardRef } from "react";
 import type { Meta, StoryObj } from "@storybook/react";
-import { useArgs } from "@storybook/client-api";
+import { useArgs } from "@storybook/preview-api";
 
 import { ChevronRightIcon, HomeIcon } from "@nimbus-ds/icons";
 import { Tag, Badge, Icon } from "@nimbus-ds/components";