+++ packages/react/src/components/SideModal/src/sideModal.stories.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import type { Meta, StoryObj } from "@storybook/react";
-import { useArgs } from "@storybook/client-api";
+import { useArgs } from "@storybook/preview-api";
 import { Icon, Text, Button, Box } from "@nimbus-ds/components";
 import {
   ChevronLeftIcon,